Version 0.2 (1.2 beta)
Complete overhaul of the Token Store and Auction House.

Token Store:
  - You are now able to set items in a `storeItems.yml` file
  - The Token Store will automatically update upon server restart/refresh
  - Fix some indexing issues

Auction House:
  - Indexing fixes

About Command:
  - Update information

Global:
  - Change `Config.java` to `Storage.java` because `Config.config`, `Config.data` and `Config.storeItems` doesn't make that much sense

+++ src/main/java/ca/cxtokens/Shop/Static/Item.java
@@ -6,9 +6,9 @@ import org.bukkit.inventory.ItemStack;
 public class Item {
     public ItemStack stack = new ItemStack(Material.AIR, 0);
     public int price = 0;
-    public float sellMultiplier = 0.5f;
+    public double sellMultiplier = 0.5;
 
-    public Item(ItemStack stack, int price, float sellMultiplier) {
+    public Item(ItemStack stack, int price, double sellMultiplier) {
         this.stack = stack;
         this.price = price;
         this.sellMultiplier = sellMultiplier;
